修正出身物品道具中文 隨機陣列的調整
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@103 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/model/Instance/L1SummonInstance.java
@@ -118,8 +118,8 @@ public class L1SummonInstance extends L1NpcInstance {
 				new SummonTimer(), SUMMON_TIME);
 
 		setMaster(master);
-		setX(master.getX() + RandomArrayList.getArray5List() - (byte) 2);
-		setY(master.getY() + RandomArrayList.getArray5List() - (byte) 2);
+		setX(master.getX() + RandomArrayList.getArray5List() - 2);
+		setY(master.getY() + RandomArrayList.getArray5List() - 2);
 		setMap(master.getMapId());
 		setHeading(5);
 		setLightSize(template.getLightSize());
